move Fieldset and ArrayInput one level up
