+++ src/pages/namespace/Explorer/Endpoint/EndpointEditor/Form/plugins/Inbound/AclForm.tsx
@@ -2,11 +2,12 @@ import {
   AclFormSchema,
   AclFormSchemaType,
 } from "../../../schema/plugins/inbound/acl";
-import { ArrayInput, Fieldset } from "../../components/FormHelper";
 import { Controller, useForm } from "react-hook-form";
 import { FC, FormEvent } from "react";
 import FormErrors, { errorsType } from "~/componentsNext/FormErrors";
 
+import { ArrayInput } from "~/pages/namespace/Explorer/components/ArrayInput";
+import { Fieldset } from "~/pages/namespace/Explorer/components/Fieldset";
 import { PluginWrapper } from "../components/Modal";
 import { useTranslation } from "react-i18next";
 import { zodResolver } from "@hookform/resolvers/zod";
