+++ src/pages/namespace/Explorer/Endpoint/EndpointEditor/Form/plugins/Outbound/JsOutboundForm.tsx
@@ -8,7 +8,7 @@ import {
 
 import { Card } from "~/design/Card";
 import Editor from "~/design/Editor";
-import { Fieldset } from "../../components/FormHelper";
+import { Fieldset } from "~/pages/namespace/Explorer/components/Fieldset";
 import { PluginWrapper } from "../components/Modal";
 import { useTheme } from "~/util/store/theme";
 import { useTranslation } from "react-i18next";
