+++ src/pages/namespace/Explorer/components/ArrayInput.tsx
@@ -1,38 +1,9 @@
-import { FC, PropsWithChildren, useState } from "react";
+import { FC, useState } from "react";
 import { Plus, X } from "lucide-react";
 
 import Button from "~/design/Button";
 import { ButtonBar } from "~/design/ButtonBar";
 import Input from "~/design/Input";
-import { twMergeClsx } from "~/util/helpers";
-
-type FieldsetProps = PropsWithChildren & {
-  label: string;
-  htmlFor?: string;
-  className?: string;
-  horizontal?: boolean;
-};
-
-export const Fieldset: FC<FieldsetProps> = ({
-  label,
-  htmlFor,
-  children,
-  className,
-  horizontal,
-}) => (
-  <fieldset
-    className={twMergeClsx(
-      "mb-2 flex gap-2",
-      className,
-      horizontal ? "flex-row-reverse items-center" : "flex-col"
-    )}
-  >
-    <label className="grow text-sm" htmlFor={htmlFor}>
-      {label}
-    </label>
-    {children}
-  </fieldset>
-);
 
 type ArrayInputProps = {
   placeholder?: string;
