+++ src/pages/namespace/Explorer/utils.ts
@@ -6,3 +6,18 @@ import { stringify as jsonToPrettyYamlStringify } from "json-to-pretty-yaml";
  */
 export const jsonToYaml = (t: Record<string, unknown>) =>
   Object.keys(t).length === 0 ? "" : jsonToPrettyYamlStringify(t);
+
+export const treatEmptyStringAsUndefined = (value: unknown) => {
+  if (value === "") {
+    return undefined;
+  }
+  return value;
+};
+
+export const treatAsNumberOrUndefined = (value: unknown) => {
+  const parsed = parseInt(`${value}`, 10);
+  if (isNaN(parsed)) {
+    return undefined;
+  }
+  return parsed;
+};
